requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
blinker==1.9.0
blis==1.3.0
cachelib==0.13.0
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
datacommons-client==2.1.6
distlib==0.4.0
Django==5.2.8
feedparser==6.0.12
filelock==3.20.1
Flask==3.1.2
flask-cors==6.0.2
Flask-Scss==0.5
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgspec==0.19.0
murmurhash==1.0.13
newsapi-python==0.2.7
newsdataapi==0.1.29
numpy==2.4.2
packaging==25.0
pandas==3.0.0
pipenv==2026.0.3
platformdirs==4.5.1
politifact==0.1
preshed==3.0.10
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.7
pyScss==1.4.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
requests==2.32.5
rich==14.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.6
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.35.4
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.5
wrapt==1.17.2
